feat: generalise instance migration logic into custom hook
